Move getRequireDeclaratorsWithProperty to modules folder
