+++ src/transforms/v2-to-v3/client-names/index.ts
@@ -1,4 +1,3 @@
 export * from "./getClientMetadataRecord";
-export * from "./getRequireDeclaratorsWithProperty";
 export * from "./getV2ClientNamesFromGlobal";
 export * from "./getV2ClientNamesRecord";
