+++ src/transforms/v2-to-v3/modules/getRequireDeclaratorsWithProperty.ts
